+++ src/Fluent.Calculations.Primitives.Tests/Arrays/ValueArrayTests.cs
@@ -1,47 +1,38 @@
 ﻿using Fluent.Calculations.Primitives.BaseTypes;
-using System.Collections;
 
-namespace Fluent.Calculations.Primitives.Tests.Arrays;
-
-public class ValueArrayTests
+namespace Fluent.Calculations.Primitives.Tests.Arrays
 {
-    [Fact]
-    public void Test()
+
+    public partial class ValueArrayTests
     {
+        [Fact]
+        public void Test()
+        {
 
 
-    }
+        }
 
-    public static class Constant
-    {
-        public const string TestEvaluationName = "TEST-ARRAY-EVALUAITON-NAME";
-    }
-    public class Values<T> : IEnumerable<IValue>
-    {
-        private readonly List<IValue> numbers = new List<IValue>();
+        public static class Constant
+        {
+            public const string TestEvaluationName = "TEST-ARRAY-EVALUAITON-NAME";
+        }
 
-        public void Add(IValue number) => numbers.Add(number);
+        public class TestArrayCalcuation
+        {
+            public static Number Return()
+            {
+                EvaluationOptions options = new() { AlwaysReadNamesFromExpressions = true };
+                EvaluationContext<Number> Calculation = new(options);
 
-        public IEnumerator<IValue> GetEnumerator() => numbers.GetEnumerator();
+                Number
+                    NumberOne = Number.Of(2),
+                    NumberTwo = Number.Of(3);
 
-        IEnumerator IEnumerable.GetEnumerator() => numbers.GetEnumerator();
-    }
+                Numbers MultipleNumbers = new() { NumberOne, NumberTwo };
 
-    //public class TestArrayCalcuation
-    //{
-    //    public static Number Return()
-    //    {
-    //        EvaluationOptions options = new() { AlwaysReadNamesFromExpressions = true };
-    //        EvaluationContext<Number> Calculation = new(options);
+                return Calculation.Evaluate(() => MultipleNumbers.Sum(), Constant.TestEvaluationName);
+            }
 
-    //        Number
-    //            NumberOne = Number.Of(2),
-    //            NumberTwo = Number.Of(3);
-
-    //        Numbers MultipleNumbers = new() { NumberOne, NumberTwo };
-
-    //        return Calculation.Evaluate(() => MultipleNumbers.Sum(), Constant.TestEvaluationName);
-    //    }
-
+        }
     }
 }
