+++ src/Fluent.Calculations.Primitives/Collections/Numbers.cs
@@ -1,0 +1,8 @@
+﻿namespace Fluent.Calculations.Primitives.Collections;
+using Fluent.Calculations.Primitives.BaseTypes;
+
+public class Numbers : Values<Number>
+{
+
+}
+
