+++ src/Fluent.Calculations.Primitives/Collections/ValuesLinqExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace Fluent.Calculations.Primitives.Collections;
+using Fluent.Calculations.Primitives.BaseTypes;
+
+public static class ValuesLinqExtensions
+{
+    public static Number Sum(this IEnumerable<IValue> values) 
+        => Number.Of(values.Sum(value => value.Primitive));
+
+}
